fix(async): close round-3 CodeRabbit findings in RequestSlot lifecycle
- completion.hpp: commit_binding_to_outstanding now CASes binding->outstanding
  and fail-fasts on both CAS losses (no silent state corruption)
- request_arena.hpp: bounds-checked introspection accessors; stale enqueue and
  generation exhaustion fail fast (review findings #4/#5)
- fake_backend.hpp/sync_backend.hpp: side-band FIFOs/staging deques removed;
  complete_oldest_* resolves via bounded slot scan + record_terminal; test-only
  sink_deliveries() and SubmitPauseGate guarded behind
  SLUICE_ASYNC_INTERNAL_TESTING
- reference_ready_sink.hpp: delivery counter is test-only
- fail_fast.{hpp,cpp}: request_arena_enqueue_stale_fail_fast,
  request_arena_generation_exhausted_fail_fast,
  request_arena_slot_index_out_of_range_fail_fast
- tests: stale-enqueue + generation-exhaustion death cases with controls,
  commit-binding-without-begin death case, aligned operator new in no-alloc
  probe, ready-ring reap-order test, exactly_one_terminal_winner asserts the
  published terminal bytes, cancel-then-reuse and double-complete regressions
- xmake/tests: reference_backend_arena_lifecycle_test moved to the
  internal-testing group (it uses the SLUICE_ASYNC_INTERNAL_TESTING-only
  sink_deliveries seam); ci.yml registers the arena negative-compile script;
  verify scripts use failures=$((failures+1)) instead of ((failures++)) to
  avoid set -e abort on the first failure
- docs: compliance-gate ABI claim corrected (Completion<T> layout DOES change),
  gate counts reconciled (134/134, negative-compile 18/18); ADR/design closeout
  notes updated

+++ include/sluice/async/detail/request_key.hpp
@@ -13,7 +13,12 @@
 //                     detail owned by the arena/context).
 //   SlotIndex       — index into the bounded RequestSlot arena.
 //   Generation      — per-slot ABA guard; incremented on slot release BEFORE the
-//                     next key can become visible (I6).
+//                     next key can become visible (I6). 64-bit so a stale key
+//                     can NEVER collide with the current generation (I6 absolute
+//                     wording): ~584 years at one release per nanosecond. The
+//                     arena fail-fasts at UINT64_MAX rather than silently wrap,
+//                     so the ABA property holds in perpetuity (review finding:
+//                     32-bit wrap re-introduces ABA under heavy reuse).
 //
 // RequestKey is a trivial value type: copyable, comparable, suitable for use as
 // a stable value identity (e.g. carried by-value in a ReadyEvent). It MUST NOT
@@ -46,7 +51,7 @@ struct SlotIndex {
 };
 
 struct Generation {
-    std::uint32_t value = 0;
+    std::uint64_t value = 0;
     friend bool operator==(const Generation&, const Generation&) noexcept = default;
 };
